Fix typo
From Litle_XML_Reference_Guide_XML8.13_V2.24.pdf

4.163 payPalOrderComplete
The payPalOrderComplete element is an optional child of both the capture and sale elements, but is required to close a PayPal order. Set the value to true to close the order, when you have fulfilled the order and do not need to send any further auths or deposits against it. Set the value to false to keep the order open for additional auths or deposits.